- Login para inscripción.
- Formulario de inscripción
- Validaciones para inscripción al programa.
- Carga de estilos y js para el funcionamiento de la página.
- Procesamiento de la inscripción en Base de Datos.
- Instalación de framework de diseño "CUBA"
- Login para CRM
- Menú dinámico según el usuario.
- Cierre de sesiones.
- Consulta de socio

+++ bepensa-socio-selecto-data/models/BitacoraDeContrasena.cs
@@ -7,7 +7,7 @@ public partial class BitacoraDeContrasena
 {
     public long Id { get; set; }
 
-    public long? IdUsuario { get; set; }
+    public int? IdUsuario { get; set; }
 
     public int? IdOperador { get; set; }
